refactor(token): 패키지 이동
